Revert "added results page"
