+++ src/App.tsx
@@ -5,8 +5,8 @@ import {BrowserRouter as Router, Route, Routes} from 'react-router-dom';
 import Home from './components/pages/Home';
 import Basic from './components/pages/Basic';
 import Detailed from './components/pages/Detailed';
-import Results from './components/pages/Results';
-import Footer from './components/Footer';
+import API from './components/pages/API';
+//import Footer from './components/Footer';
 import Header from './components/Header'
 
 
@@ -21,12 +21,11 @@ function App() {
         <Route path="/" element={<Home />} />
         <Route path="/basic" element={<Basic />} />
         <Route path="/detailed" element={<Detailed />} />
-        <Route path="/results" element={<Results />} />
+        <Route path="/api" element={<API />} />
         </Routes>
       </main>
     </div>
     </Router>
-    <Footer />
     </div>
   );
 }
