+++ src/components/Footer.tsx
@@ -1,40 +1,25 @@
-import React from 'react'
-import { useState } from 'react';
-import {Form} from 'react-bootstrap'
-import {Button} from '@mui/material';
+import React from 'react';
 
-import './styles/Footer.css'
-
-//local storage and API Key: key should be entered in by the user and will be stored in local storage (NOT session storage)
-let keyData = "";
-const saveKeyData = "MYKEY";
-const prevKey = localStorage.getItem(saveKeyData); //so it'll look like: MYKEY: <api_key_value here> in the local storage when you inspect
-if (prevKey !== null) {
-  keyData = JSON.parse(prevKey);
-}
 const Footer: React.FC = () => {
-    const [key, setKey] = useState<string>(keyData); //for api key input
-  
-  //sets the local storage item to the api key the user inputed
-  function handleSubmit() {
-    localStorage.setItem(saveKeyData, JSON.stringify(key));
-    window.location.reload(); //when making a mistake and changing the key again, I found that I have to reload the whole site before openai refreshes what it has stores for the local storage variable
-  }
-
-  //whenever there's a change it'll store the api key in a local state called key but it won't be set in the local storage until the user clicks the submit button
-  function changeKey(event: React.ChangeEvent<HTMLInputElement>) {
-    setKey(event.target.value);
-  }
     return (
-        <div className='Footer'>
-      <Form className='Footer-API'>
-        <Form.Label></Form.Label>
-        <Form.Control style = {{width: '50%'}}type="password" placeholder="Insert API Key Here" onChange={changeKey}></Form.Control>
-        <br></br>
-        <Button variant = 'contained' sx = {{backgroundColor: '#EF233C'}} onClick={handleSubmit}>Submit</Button>
-      </Form>
-      </div>
+        <footer style={styles.footer}>
+            <h1></h1>
+            <div>
+                <p>Contributors: Jonathan Liss, Ronald Kouloun, Mason Davis, Ronaldo Castillo</p>
+            </div>
+
+        </footer>
     )
+}
+
+const styles = {
+    footer: {
+        backgroundColor: '#575e67',
+        position: 'relative' as const,
+        bottom: 0,
+        width: '100%',
+        height: 80,
+    }
 }
 
 export default Footer;
