+++ src/components/Header.tsx
@@ -1,5 +1,6 @@
 import React from 'react'
 import {Link} from 'react-router-dom'
+import API from './pages/API';
 import './styles/Header.css';
 
 const Header: React.FC = () => {
@@ -14,7 +15,7 @@ const Header: React.FC = () => {
         <span>|</span>
         <Link to='/detailed'>Detailed Quiz</Link>
         <span>|</span>
-        <Link to='/results'>Results</Link>
+        <Link to='/api'>API</Link>
     </nav>
         </header>
         </div>
