--- src/components/pages/Results.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Results: React.FC = () => {
-    return (
-    <div>
-        <h1>Career Results</h1>
-    </div>
-    )
-}
-
-export default Results;
