+++ src/components/styles/API.css
@@ -1,13 +1,11 @@
-.Footer {
+.API {
     text-align: center;
 }
 
-.Footer-API {
+.API-bar {
     display: flex;
     justify-content: center;
     place-items: center;
-    padding: 60px;
+    padding: 20px;
     white-space: nowrap;
-    position: bottom;
-    min-height: 40vh;
 }
